appending

+++ Dubbo/src/main/java/com/hzgc/dubbo/dynamicrepo/CapturedPicture.java
@@ -24,7 +24,7 @@ public class CapturedPicture implements Serializable {
     /**
      * 图片的相似度
      */
-    private int similarity;
+    private Float similarity;
     /**
      * 图片的附加信息，扩展预留
      */
@@ -66,11 +66,11 @@ public class CapturedPicture implements Serializable {
         this.description = description;
     }
 
-    public int getSimilarity() {
+    public Float getSimilarity() {
         return similarity;
     }
 
-    public void setSimilarity(int similarity) {
+    public void setSimilarity(Float similarity) {
         this.similarity = similarity;
     }
